Integrate new Cache backend into ContainerBuilder
Instead of tracking the caching status with a boolean, this commit changes
`$sourceCache` to a nullable string. This allows to define different cache-
backends.

`::enableDefinitionCache(..)` not takes an optional parameter - optional to
keep backwards compatibility - with the class-name of the definition cache to
use. By default the `ApcuCache` is used which is identical to the deprecated,
original `SourceCache`