Attempt to fix #648 and keep the container idempotent
The idea is that each definition is compiled into a method which bears a unique name. That unique name is a hash generated from the definition.

Each definition generates its own hash, so that two different definitions have two different hashes. But it's okay if different definitions have the same hash if they are actually equal.