Make `injectOn` use the definition cache
Before this change `injectOn` was not using the Container internal
definition cache and insetad solely built on the optional SourceCache
mechanism. This is costly for setups with a long-running processes with
APCu as every `injectOn` call will lead to a full annotation and
autowiring run.

This change makes `injectOn` use the definition cache for all objects of
non-anonymous classes. Anonymmous objects are ignored to make sure the
cache does not get overloaded with many unusable definitions. It also
includes a basic test which tests for minimum compatibility of `injectOn`
with anonymous classes.