Add functionality to reset container entries
The purpose of this change is to improve the support for resource
management and introspection of the `Container` class by adding the
following methods:

* Container::reset - unsets an entry in the container

* Container::initialized - checks if an entry is initialized

These method names were chosen for consistency with the same features
in Symfony's Dependency Injection component.

The idea here is that if you have an entry in the container that is no
longer needed (say, for example, a database connection), then you can
remove the entry from the container with `reset` to free the resource.
If it is needed again later, no extra work is needed to `get` it from
the container according to its definition, and a new instance will be
constructed.

The `initialized` method is related in the sense that there is no
existing way to see if an entry is present in the container. It is
distinct from the `has` method, which only reports if an entry _can_
be supplied by the container, not if it _has_. You may have some logic
that depends on if a resource has been allocated yet (i.e. before the
first `get` or after `reset`).