ContainerBuilder: template on container class
This is a change to docstrings only.

The type of container that is actually built by the ContainerBuilder
can be more specialized than the Container base class. This can occur
in two scenarios:

1. Using the `$containerClass` argument of the constructor. Note that
   this class must still inherit from Container.

2. Using the `$containerParentClass` argument of `enableCompilation`.
   Note that this class must still inherit from CompiledContainer.

Below is an enumeration of all possible template configurations:

  new ContainerBuilder(); // ContainerBuilder<Container>
  new ContainerBuilder(MyContainer::class); // ContainerBuilder<MyContainer>
  $builder->enableCompilation(...); // ContainerBuilder<CompiledContainer>
  $builder->enableCompilation(..., MyCompiledContainer::class); // ContainerBuilder<MyCompiledContainer>

This should fix many static analysis errors for users that specify a
custom container class. Note that when the default arguments are used,
the original behavior is recovered.