ContainerBuilder: fallback typehints for PhpStorm
Fixes #844.

PhpStorm currently lacks the ability to infer a template type from a
`class-string` constructor argument (see
https://youtrack.jetbrains.com/issue/WI-65811/Infer-type-from-constructors-class-string-parameter).
As a result of #842 (v7.0.1), this requires PhpStorm users to manually
typehint their ContainerBuilder instances, e.g.:

    /** @var ContainerBuilder<Container> $builder */
    $builder = new ContainerBuilder();

To avoid this, we only use the proper templated typehints with Psalm
annotations (which will be usable by both Psalm and PHPStan), and leave
the non-templated typehints as a fallback. Note that the fallbacks will
not give the correct type when using a non-default Container class.